fix(wasmJs): use toDouble() before toLong() for JsNumber conversion
JsNumber.toLong() doesn't exist in wasmJs target. Convert through
Double first, matching the pattern used in the js target.

+++ cmp-paycraft/src/wasmJsMain/kotlin/com/mobilebytelabs/paycraft/platform/TimeProvider.wasmJs.kt
@@ -5,4 +5,4 @@ import kotlin.js.JsNumber
 @JsFun("() => Date.now()")
 private external fun dateNow(): JsNumber
 
-actual fun currentTimeMillis(): Long = dateNow().toLong()
+actual fun currentTimeMillis(): Long = dateNow().toDouble().toLong()
